Issue #154 refactor: TimerService / TimeConverter
