+++ app/src/main/java/com/lateinit/rightweight/service/TimerService.kt
@@ -13,17 +13,10 @@ import androidx.core.app.NotificationCompat
 import androidx.lifecycle.LifecycleService
 import androidx.lifecycle.lifecycleScope
 import com.lateinit.rightweight.R
+import com.lateinit.rightweight.util.convertTimeStamp
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-
-fun convertTimeStamp(count: Int): String {
-    val hours: Int = (count / 60) / 60
-    val minutes: Int = (count / 60) % 60
-    val seconds: Int = count % 60
-
-    return "${"%02d".format(hours)}:${"%02d".format(minutes)}:${"%02d".format(seconds)}"
-}
 
 class TimerService : LifecycleService() {
     private var isTimerRunning = false
