+++ app/src/main/java/com/lateinit/rightweight/ui/exercise/ExerciseFragment.kt
@@ -25,13 +25,13 @@ import com.lateinit.rightweight.service.TimerService.Companion.STATUS
 import com.lateinit.rightweight.service.TimerService.Companion.STATUS_ACTION_NAME
 import com.lateinit.rightweight.service.TimerService.Companion.STOP
 import com.lateinit.rightweight.service.TimerService.Companion.TIME_COUNT_INTENT_EXTRA
-import com.lateinit.rightweight.service.convertTimeStamp
 import com.lateinit.rightweight.ui.dialog.CommonDialogFragment
 import com.lateinit.rightweight.ui.dialog.CommonDialogFragment.Companion.END_EXERCISE_DIALOG_TAG
 import com.lateinit.rightweight.ui.model.ExercisePartTypeUiModel
 import com.lateinit.rightweight.ui.model.HistoryExerciseSetUiModel
 import com.lateinit.rightweight.ui.model.HistoryExerciseUiModel
 import com.lateinit.rightweight.util.collectOnLifecycle
+import com.lateinit.rightweight.util.convertTimeStamp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
